Amendment: Election and regulation of joint committee positions (#1)
## Rationale

The Constitution currently does not have anything on joint holding of Committee positions. This means that they’re technically not forbidden but they’re also undefined behaviour. This proposal explicitly allows them and regulates how they work.

## Changes

Does not cover what happens in the situation of dismissal - no disagreements on simply including dismissal in the same clause that deals with resignation.

## Vote

Final vote: 21 for (91%), 0 against, 2 abstaining
Passed (with changes)

Co-Authored-By: Ash Holland <ash@sorrel.sh>