Amendment: Make it harder to dissolve the society (#3)
## Rationale

It’s ludicrous that dissolving the society only takes a simple majority vote when amending the constitution requires a two-thirds majority. This proposal switches the majorities required.

## Changes

No issues with upping the majority required for dissolving the society but concerns were raised over lowering the majority for an amendment. Attendees voted to remove the change to the majority required for an amendment.

## Vote

Final vote: 22 for (96%), 1 against, 0 abstaining
Passed (with changes)