Amendment: Appendices (#8)
## Rationale

The Constitution regularly leaves out details on matters.

For example, §10.5.2 specifies that "Single Transferable Vote" should be used to count votes but STV is not a single method of counting votes. There are many variations that can produce significantly different results. As it stands, this is undefined and largely ends up being down to the Returning Officer to decide.

It makes sense for such specificities to be beyond the scope of the constitution. Such details are too important to leave to in-the-moment interpretation while not important enough that they should require going to the lengths of organising a General Meeting and getting a two-thirds majority to amend. They can also be notably complex.

To fix this, this proposal amends the constitution to allow "appendices". These constitutional appendices are required to be specified by the constitution and limited in scope, and they are able to be amended by a simple majority of the committee. These measures respectively prevent appendices from effectively being a backdoor to amending the constitution while maintaining reasonable democratic control over them.

## Vote

Final vote: 16 for (94%), 0 against, 1 abstaining
Passed