Renaming committee positions
This amendment seeks to reduce the confusion between the Secretary and Social Secretary roles, as well as Ordinary Members, by:
- Removing the reference to the 'Society Secretary' in section 7.2 and replacing it with 'Secretary', like the other signatory roles within that section;
- Renaming Social Secretary to Social Events Officer, to bring it in line with other non signatory positions such as Academic Events Officer;
- Renaming Ordinary Member to Ordinary Officer, to bring it in line with other non signatory positions.

Currently, it could be assumed that the Secretary and the Social Secretary were somehow linked, or the two could be easily confused. For Ordinary Members, it could potentially be confusing whether someone was referring to an Ordinary Member, or a member of the society that did not hold a place on the committee.